# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RecruitIQ API — recruitment knowledge management system
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  ok: boolean;
}

export interface AuthUser {
  id: number;
  name: string;
  email: string;
}

export interface LoginInput {
  email: string;
}

/**
 * @nullable
 */
export type CandidateSource = typeof CandidateSource[keyof typeof CandidateSource] | null;


export const CandidateSource = {
  resume: 'resume',
  chat: 'chat',
  voice: 'voice',
  mixed: 'mixed',
} as const;

export interface Candidate {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  resumeText?: string | null;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  source?: CandidateSource;
  createdAt: string;
  updatedAt: string;
}

export interface CandidatePreference {
  id: number;
  candidateId: number;
  /** @nullable */
  desiredRole?: string | null;
  /** @nullable */
  desiredRoleCategory?: string | null;
  /** @nullable */
  salaryExpectation?: string | null;
  /** @nullable */
  noticePeriod?: string | null;
  /** @nullable */
  wfhPreference?: string | null;
  /** @nullable */
  relocationWillingness?: boolean | null;
  /** @nullable */
  shiftPreference?: string | null;
  /** @nullable */
  careerInterests?: string | null;
  /** @nullable */
  lastVerifiedAt?: string | null;
  updatedAt: string;
}

export interface CandidateWithPreferences {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  resumeText?: string | null;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  source?: string | null;
  createdAt: string;
  updatedAt: string;
  preferences?: CandidatePreference;
}

export interface CandidatePreferenceInput {
  desiredRole?: string;
  desiredRoleCategory?: string;
  salaryExpectation?: string;
  noticePeriod?: string;
  wfhPreference?: string;
  relocationWillingness?: boolean;
  shiftPreference?: string;
  careerInterests?: string;
  lastVerifiedAt?: string;
}

export interface CandidateInput {
  name: string;
  email?: string;
  phone?: string;
  resumeText?: string;
  skills?: string;
  source?: string;
  preferences?: CandidatePreferenceInput;
}

export interface CandidateUpdate {
  name?: string;
  email?: string;
  phone?: string;
  resumeText?: string;
  skills?: string;
  source?: string;
  preferences?: CandidatePreferenceInput;
}

export interface DuplicateCheckInput {
  email?: string;
  phone?: string;
}

export interface DuplicateCheckResult {
  isDuplicate: boolean;
  existingCandidate?: Candidate;
}

export interface ExtractedCandidateInfo {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  skills?: string | null;
  /** @nullable */
  desiredRole?: string | null;
  /** @nullable */
  desiredRoleCategory?: string | null;
  /** @nullable */
  salaryExpectation?: string | null;
  /** @nullable */
  noticePeriod?: string | null;
  /** @nullable */
  wfhPreference?: string | null;
  /** @nullable */
  relocationWillingness?: boolean | null;
  /** @nullable */
  shiftPreference?: string | null;
  /** @nullable */
  careerInterests?: string | null;
}

export interface JobOpening {
  id: number;
  title: string;
  /** @nullable */
  roleCategory?: string | null;
  /** @nullable */
  requirements?: string | null;
  createdAt: string;
}

export interface JobOpeningInput {
  title: string;
  roleCategory?: string;
  requirements?: string;
}

export interface JobOpeningUpdate {
  title?: string;
  roleCategory?: string;
  requirements?: string;
}

export type CandidateJobStatusStatus = typeof CandidateJobStatusStatus[keyof typeof CandidateJobStatusStatus];


export const CandidateJobStatusStatus = {
  matched: 'matched',
  rejected: 'rejected',
  open_for_future: 'open_for_future',
  interested: 'interested',
  contacted: 'contacted',
} as const;

export interface CandidateJobStatus {
  id: number;
  candidateId: number;
  jobId: number;
  status: CandidateJobStatusStatus;
  /** @nullable */
  updatedBy?: string | null;
  updatedAt: string;
}

export interface CandidateMatchResult {
  candidate: CandidateWithPreferences;
  jobStatus?: CandidateJobStatus;
}

export type CandidateJobStatusWithJobStatus = typeof CandidateJobStatusWithJobStatus[keyof typeof CandidateJobStatusWithJobStatus];


export const CandidateJobStatusWithJobStatus = {
  matched: 'matched',
  rejected: 'rejected',
  open_for_future: 'open_for_future',
  interested: 'interested',
  contacted: 'contacted',
} as const;

export interface CandidateJobStatusWithJob {
  id: number;
  candidateId: number;
  jobId: number;
  status: CandidateJobStatusWithJobStatus;
  /** @nullable */
  updatedBy?: string | null;
  updatedAt: string;
  job?: JobOpening;
}

export type CandidateJobStatusInputStatus = typeof CandidateJobStatusInputStatus[keyof typeof CandidateJobStatusInputStatus];


export const CandidateJobStatusInputStatus = {
  matched: 'matched',
  rejected: 'rejected',
  open_for_future: 'open_for_future',
  interested: 'interested',
  contacted: 'contacted',
} as const;

export interface CandidateJobStatusInput {
  candidateId: number;
  jobId: number;
  status: CandidateJobStatusInputStatus;
  updatedBy?: string;
}

export type CandidateJobStatusUpdateStatus = typeof CandidateJobStatusUpdateStatus[keyof typeof CandidateJobStatusUpdateStatus];


export const CandidateJobStatusUpdateStatus = {
  matched: 'matched',
  rejected: 'rejected',
  open_for_future: 'open_for_future',
  interested: 'interested',
  contacted: 'contacted',
} as const;

export interface CandidateJobStatusUpdate {
  status: CandidateJobStatusUpdateStatus;
  updatedBy?: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface RoleCategoryCount {
  roleCategory: string;
  count: number;
}

export interface DashboardStats {
  totalCandidates: number;
  totalJobs: number;
  byCandidateStatus: StatusCount[];
  byCandidateRoleCategory: RoleCategoryCount[];
  recentCandidates: CandidateWithPreferences[];
}

export type ListCandidatesParams = {
roleCategoryFilter?: string;
statusFilter?: string;
search?: string;
};

export type ListCandidateJobStatusesParams = {
jobId?: number;
candidateId?: number;
statusFilter?: string;
};