# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RecruitIQ API — recruitment knowledge management system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email (simple, no password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /candidate-job-status:
    get:
      operationId: listCandidateJobStatuses
      tags: [status]
      summary: List candidate job statuses with optional filters
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
        - name: candidateId
          in: query
          schema:
            type: integer
        - name: statusFilter
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CandidateJobStatus"

    post:
      operationId: upsertCandidateJobStatus
      tags: [status]
      summary: Create or update a candidate's status for a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateJobStatusInput"
      responses:
        "200":
          description: Upserted status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateJobStatus"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /candidate-job-status/{id}:
    patch:
      operationId: updateCandidateJobStatus
      tags: [status]
      summary: Update an existing candidate job status entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateJobStatusUpdate"
      responses:
        "200":
          description: Updated status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateJobStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AuthUser:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
    LoginInput:
      type: object
      required: [email]
      properties:
        email:
          type: string

    # ── Candidates ────────────────────────────────────────────────────────
...
    CandidateJobStatus:
      type: object
      required: [id, candidateId, jobId, status, updatedAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        jobId:
          type: integer
        status:
          type: string
          enum: [matched, rejected, open_for_future, interested, contacted]
        updatedBy:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
...
    CandidateJobStatusInput:
      type: object
      required: [candidateId, jobId, status]
      properties:
        candidateId:
          type: integer
        jobId:
          type: integer
        status:
          type: string
          enum: [matched, rejected, open_for_future, interested, contacted]
        updatedBy:
          type: string
    CandidateJobStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [matched, rejected, open_for_future, interested, contacted]
        updatedBy:
          type: string

    # ── Dashboard ─────────────────────────────────────────────────────────
...