# lib/api-zod/src/generated/types/candidateJobStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RecruitIQ API — recruitment knowledge management system
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateJobStatusStatus } from './candidateJobStatusStatus';

export interface CandidateJobStatus {
  id: number;
  candidateId: number;
  jobId: number;
  status: CandidateJobStatusStatus;
  /** @nullable */
  updatedBy?: string | null;
  updatedAt: Date;
}