# lib/api-zod/src/generated/types/candidateJobStatusInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RecruitIQ API — recruitment knowledge management system
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateJobStatusInputStatus } from './candidateJobStatusInputStatus';

export interface CandidateJobStatusInput {
  candidateId: number;
  jobId: number;
  status: CandidateJobStatusInputStatus;
  updatedBy?: string;
}