# lib/api-zod/src/generated/types/candidateJobStatusUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * RecruitIQ API — recruitment knowledge management system
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateJobStatusUpdateStatus } from './candidateJobStatusUpdateStatus';

export interface CandidateJobStatusUpdate {
  status: CandidateJobStatusUpdateStatus;
  updatedBy?: string;
}